@@ -1,6 +1,6 @@
 {
   "name": "dts2as",
-  "version": "0.11.0",
+  "version": "0.12.0",
   "description": "A command line tool that converts TypeScript definitions (d.ts files) to ActionScript 3 classes and interfaces",
   "repository": "BowlerHatLLC/dts2as",
   "author": "Josh Tynjala",
@@ -22,15 +22,20 @@
     "prepublish": "npm run test"
   },
   "dependencies": {
-    "minimist": "^1.1.3",
+    "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
-    "rimraf": "^2.4.3",
-    "typescript": "^1.8.7"
+    "rimraf": "^2.6.1",
+    "typescript": "^2.2.1"
   },
   "devDependencies": {
-    "del": "^2.1.0",
-    "gulp": "^3.9.0",
-    "gulp-jasmine": "^2.2.1",
-    "gulp-typescript": "^2.12.1"
+    "@types/jasmine": "^2.5.46",
+    "@types/minimist": "^1.2.0",
+    "@types/mkdirp": "^0.3.29",
+    "@types/node": "^7.0.11",
+    "@types/rimraf": "0.0.28",
+    "del": "^2.2.2",
+    "gulp": "^3.9.1",
+    "gulp-jasmine": "^2.4.2",
+    "gulp-typescript": "^3.1.6"
   }
 }
